Add a new command line option (`-sort-free-variable-strategy`) that allows the sort order of free variables to be controlled.
This a hack. It appears that for benchmarks with a large number of
free variables doing the sort is incredbily expensive so this option
provides a way of avoiding it for now. We need a proper strategy for
this though. We need determinisim but also a strategy that is cheap.