Rework runtime assertions. Previously we were just using C's `assert()` macro which was inappropriate because our fuzzing target is currently a call to `abort()` which is the same thing that happens when `assert()` fails. This would mean if the asserts failed we would treat that as if we hit the fuzzing target even though we didn't.
All runtime asserts (and calls to abort()) have been replaced with
the new `jassert()` macro (`JFS_RUNTIME_FAIL()` macro). These asserts
are off by default when using the `jfs` tool but can be enabled by the
new `-runtime-asserts` command line option.

The `jassert()` macro calls the `JFS_RUNTIME_FAIL()` macro which by
default calls exit with an exit code that is distinct from the exit
code we tell LibFuzzer to use when it's found the target. This way
we now report a query as `unknown` if there's a runtime failure.

The unit tests by default build with runtime asserts enabled and
they also continue to use `abort()` when a runtime assert fails.