[CMake] Large refactor of runtime and run time unit tests.
The motivation behind this change is to start moving the implementation
of runtime functions out of the header files and into `.cpp` files
so we can compile them beforehand to try and make invocations of
clang by JFS faster.

Currently no implementations have been moved over yet and runtime header
file location is still in flux.

No modifications have been made to the JFS libraries yet to pick up
the runtime library.

This work required a large reorganisation of how runtime unittests.
Now runtime unittests are part of an external project that also builds
the runtime. The external project runs the unittests and we expose a
new top-level target `unittests` that will trigger running all the
unit tests. These tests consume the gtest build of the main project
so we don't have to build GTest multiple times.

We now support multiple verisons of runtime. Currently we have
a "release with debug symbols" and "release with debug symbols + ASan +
UBSan". We still need to teach the JFS libraries about these runtimes.