Move the implementation of `makeBitVectorFrom()` out of the header file.
This requires using a different interface being changed because
`std::enable_if<>` doesn't seem to work when trying to have a
declaration in a header and definition in an implementation file.

The aim is to slowly move the implementation of `BitVector<>` over
to a C interface so that we can use the implementation from LLVM's
JIT in the future.

The existing `BitVector<>` templates will just call the C interface so
we can avoid making changes CXXProgramBuilder.