Fix race condition in ScopedTimerImpl that would cause a bounded deadlock.
If ScopedTimerImpl destructor is called before the waiting thread
has managed to initialise (i.e. start waiting on the condition variable)
then when the notification is sent it will get missed and so the
thread calling the ScopedTimerImpl destructor will deadlock.

This is a bounded deadlock because the waiting thread waits for
a finite amount of time.

The fix is basically to spin in the ScopedTimerImpl destructor until
the waiting thread has started waiting. We might take a performance
hit for doing this but at least it fixes the deadlock.