Implement CXXProgram and CXXProgramBuilderPassImpl side of tracking the number of constraints solved.
A new library `StatLog` has been added which will record statistics at
runtime and write them to a file. This library hasn't been implemented
yet so it just contains stubs.

This implementation relies on a not yet implemented feature in
LibFuzzer. We emit a special function named `LLVMFuzzerAtExit` which
will be called by LibFuzzer just before it exits.