Patch LibFuzzer to fix fundemental problem in the way mutations are created.
For our use case the input buffer has to be an exact size and should
not be shrunk or grown. However LibFuzzer's default mutators which
we use try to grow and shrink the input.

To workaround this a new command line option has been added to
LibFuzzer (`default_mutators_resize_input=1`). In LibFuzzer itself
the option is on by default (which was the upstream behaviour would
need to be) but for JFS we set it to false by default so that the mutations
that would resize the input are disabled.

This behaviour is controllable from `jfs` itself by the new
`-libfuzzer-default-mutations-resize-input` command line option.