Implement support for customising the byte alignment of `BufferElement`s.
A lot of plumbing has been added to propagate this option through
various classes. This behaviour is exposed in the command line
interface using the `-byte-aligned-buffer-elements` option which
is false by default (buffer elements are tightly packed with no
alignment). If the option is true every `BufferElement` is byte
aligned.

It is unlikely that using byte alignment will add any benefit right
now because:

* The runtime functions for constructing free variables from the
buffer are not optimised for the byte aligned case.
* The current free variable sorting strategy (first observed) will
lead to lots of wasted bits.
* The LibFuzzer mutators are not aware of the structure of the data
and so will likely mutate padding bits. This can be fixed if we
implement our own mutator functions.