Stub out pure random fuzzer
Adds a basic entry point for a pure random fuzzer that will try random values
for use as a control group to compare against the coverage guided LibFuzzer
version.

Along with an empty `main` for the fuzzer, this wires up a new JFS option
`-libfuzzer-pure-random` to switch to this mode and adds the relevant build
scripts.