Export all symbols on Linux
On Linux, symbols inside the main executable are not exported in the symbol
table by default.  By adding the `--export-dynamic` linker flag, we can make the
available.  This allow fuzzer implementations to use a common technique to find
optional functions like `LLVMFuzzerAtExit` for both Linux and macOS.  (On macOS,
such symbols are exported by default.)