For try-all and try-all-imncsf branch encodings emit code to update the "maximum number of satisfied constraints" counter after all constraints have been evaluated and not after every constraint that evaluates as true.
The reason for doing this is because previously each chunk of counter
update code (which we emitted in every "true block" for a constraint)
counted as a new branch. This meant increases in the maximum number of
satisfied constraints counted as new coverage in the `try-all` encoding
when `-record-max-num-satisfied-constraints` was passed. This is not
ideal because this is what the `try-all-imncsf` encoding is trying to
investigate.

This change mitigates the issue because the fuzzer only
gets feed back for increasing the "maximum number of constraints"
satisfied once (the first time any constraint gets satisfied).
During this first time there's extra coverage on the the "true"
part of the constraint branch so this input would be added to the corpus
anyway.