Change many vars error in `jfs-smt2cnf` to warning
It's possible to have more CNF variables than fuzzing bits,
so we shouldn't treat that as an error.

Here's the Context of the discussion.

Subject: jfs-smt2cnf uses more variables in CNF than fuzzing bits

---
jryans:

For some benchmarks, such as QF_FP/griggio/fmcad12/sqrt.c.5.smt2,
jfs-smt2cnf will end up outputting more boolean variables in CNF form
than bits in the JFS fuzzing buffer. For this benchmark, we get 544 bits
in the fuzzing buffer and 608 SAT variables in CNF form.

I am not quite sure how to investigate what's happening here... The
benchmark expands to a very large problem in CNF form (210 MB file). I
could try remove portions of the benchmark until the discrepancy
disappears in an attempt to hunt for the root cause, I suppose.

@delcypher, should we accept such results as a possibility and model
count them as-is, or try to resolve this issue first?
---
jryans:

For a bit of clarification, this case is so far relatively rare. For
most benchmarks, the SAT variables and fuzzing bits are equal. In a few
cases, there are more fuzzing bits than variables due to padding in the
JFS buffer. The benchmark seems to be first where there are less fuzzing
bits than variables.
---
delcypher:

@jryans Sorry for being slow.

It's not entirely unexpected that the number of boolean variables in the
generated SAT problem is more than the number of fuzzing bits. The
[tseytin transform](https://en.wikipedia.org/wiki/Tseytin_transformation) that's used to convert to CNF form may introduce extra
variables. Operations that introduce some form of non-determinism could
also add extra variables.

One way to understand what's going on is to look at the "model
converter" which describes how to convert models as they propagated
through various tactics. To see this you can replace `(check-sat)` in a
benchmark with `(apply (and-then fpa2bv simplify bit-blast)
:print_benchmark true :print false :print_model_converter true)`.
Unfortunately such an approach is really only good for very small
examples. The particular example you give generates a huge amount of
output that is not digestible for humans.

My advise would be for us to ignore this problem for now. Although it is
interesting we risk getting bogged down in the weeds of precisely Z3 is
implemented. We can always come back to this later.
---
jryans:

Thanks for the advice! I hadn't seen the model converter option yet, so
that's good to know about.

Since it's not that unexpected, I will update jfs-smt2cnf to remove the
failure case when this happens then.
---