Introduce static methods to generate some special floating point constants.
- absolute largest/smallest positive/negative subnormal.
- aboslute largest/smallest positive/negative subnormal.

Some unit tests have been added to make sure these work properly.  The
underyling Z3 API that we use (`Z3_mk_fpa_numeral_int64_uint64(..)`) is
quite confusing and so comments are included that try to explain what is
going on.

This will be useful in the future for generating "smart seeds".